requirements.txt
fastapi
uvicorn

asyncio
sqlalchemy
alembic
psycopg2-binary
asyncpg
greenlet
pgvector

python-multipart

python-dotenv

openai
open-clip-torch
torch
Pillow
boto3

langgraph
langchain
langchain-openai
langchain-core
pydantic
langgraph-checkpoint-postgres

python-slugify
umap-learn
scikit-learn
plotly
pandas
hdbscan

newspaper3k
lxml-html-clean